90% done except one error
